Don't enable bash completion inside MacVim
MacVim.app runs a login shell and executes the app inside it to ensure
that all the environment variables are set correctly. To quicken
startup, this disables bash command and alias completion when it detects
that it is being run by MacVim.app. At present, this can only be
detected heuristically, and the heuristic I have settled on is to check
that the shell level is 1 and that the TERM environment variable is
unset.